Inventing is a combination of brains and materials. The more brains you use, the less material you need

+++ 0x01-variables_if_else_while/100-print_comb3.c
@@ -10,7 +10,7 @@ int main(void)
 	int i;
 	int j;
 	int find;
-	int c;
+	long unsigned int c;
 	int list[100];
 
 	for (i = 0; i <= 9; i++)
